requirements.txt
Django
psycopg2-binary==2.8.5
Pillow==7.1
gunicorn
django-modeltranslation